+++ lib/kitsu_methods.dart
@@ -6,4 +6,5 @@
 /// It contains and group up all the methods.
 library kitsu_methods;
 
-export 'package:kitsumon/src/methods/authentication_methods.dart';
+export 'package:kitsumon/src/methods/authentication_method.dart';
+export 'package:kitsumon/src/methods/characters_and_people.dart';