+++ lib/src/kitsumon.dart
@@ -4,6 +4,7 @@
  */
 
 import 'package:kitsumon/src/core/kitsu.dart';
+import 'package:kitsumon/src/methods/characters_and_people.dart';
 import 'package:meta/meta.dart';
 
 /// This connects all the various methods [Kitsu] can provide.
@@ -16,8 +17,11 @@ class Kitsumon {
   /// The [Kitsu] object. Helps creating each methods.
   final Kitsu _api;
 
+  CharactersAndPeople _charactersAndPeople;
+
   Kitsumon._(this._api) {
     _instance = this;
+    _charactersAndPeople = CharactersAndPeople(_api);
   }
 
   /// It creates the Kitsumon instance.
@@ -25,4 +29,6 @@ class Kitsumon {
 
   /// It returns the Api associated with this instance of Kitsumon.
   Kitsu get api => _api;
+
+  CharactersAndPeople get charactersAndPeople => _charactersAndPeople;
 }