+++ CHANGELOG.md
@@ -1,16 +1,43 @@
-## 0.0.2
-#### Base Implementation of Kitsumon.
+## 0.0.3
+### Still working at the base implementation of Kitsumon
 
-- Created Kitsu object.
-- Created API and Kitsumon exceptions.
-- Created KitsuValueNormalizer helper.
-- Created Authentication Method Class. (The only available one, more in the future?)
-- Created KitsuClient.
-- Created first KitsuObject, Authentication.
-- Created base implementation of Kitsumon.
-- Added build and clean Scripts.
-- Updated the README.md after talking with Kitsu developers.
+  - Fixing the **README.md**.
+  - Created **AnimeCharacterRole** enum.
+  - Created **AnimeProductionRole** enum.
+  - Simplified the **AuthenticationMethod**.
+  - Created **BaseMethod** for other methods.
+  - Initial implementation of **AnimeCharactersMethods**.
+      > *Lots of work* to do such as stating thinking the stucture.
+  - Created **CharactersAndPeople** *group* of methods.
+  - Created **Filter** object.
+  - Created **Includes** object.
+  - Fixed **Kitsu** object after *AuthenticationMethod* simplification.
+  - Fixed **KitsuClient** object to obtain only the *first error*.
+  - Added new **AnimeCharactersMethods** on methods' library.
+  - Added a direct caller on Kitsumon for **AnimeCharactersMethods**.
+  - Started writing an example to help find:
+      > *The easiest way* to use this wrapper.
+      > Errors while writing the wrapper.
+  - Fixed* Kitsumon exceptions* library name.
+  - Fixed **Pagination** object for better interaction with **Request** object.
+  - Created **Request** object that helps to route different type of API request methods.
+      > For now is implemented a starting point for *GET* *POST* *AUTH*
+  - Fixed **Sorting** object for better interaction with **Request** object.
+  - Created **SparseFieldSets** object that helps in this kind of *filter*.
+
+## 0.0.2
+### Base Implementation of Kitsumon
+
+  - Created Kitsu object.
+  - Created API and Kitsumon exceptions.
+  - Created KitsuValueNormalizer helper.
+  - Created Authentication Method Class. (The only available one, more in the future?)
+  - Created KitsuClient.
+  - Created first KitsuObject, Authentication.
+  - Created base implementation of Kitsumon.
+  - Added build and clean Scripts.
+  - Updated the README.md after talking with Kitsu developers.
 
 ## 0.0.1
 
-- Initial version. Created the Package's repo structure.
+  - Initial version. Created the Package's repo structure.